rename app accounts to users and update app users
