+++ config/urls.py
@@ -21,7 +21,7 @@ from django.conf import settings
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('allauth.urls')),
-    path('accounts/', include('accounts.urls')),
+    path('users/', include('users.urls')),
     path('', include('pages.urls')),
     path('books/', include('books.urls')),
     path('cart/', include('cart.urls'))
