+++ pages/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from accounts.models import User
+from users.models import User
 
 from django.db import models
