+++ templates/_base.html
@@ -53,7 +53,7 @@
           <a class="p-2 text-dark" href="{% url 'about' %}">About</a>
           {% if user.is_authenticated %}
             <a class="p-2 text-dark" href="{% url 'account_logout' %}">Log Out</a>
-            <a class="btn btn-outline-primary" href="{% url 'profile' %}">{{user}}</a>
+            <a class="btn btn-outline-primary" href="{% url 'users:profile' %}">{{user}}</a>
           {% else %}
               <a class="p-2 text-dark" href="{% url 'account_login' %}">Log In</a>
               <a class="btn btn-outline-primary" href="{% url 'account_signup' %}">Sign Up</a>
