+++ templates/profile.html
@@ -166,7 +166,7 @@
                       <hr><br>
                       <div class="row">
                         <div class="col-sm-12">
-                          <a class="btn btn-info " href="{% url 'update-profile' user.profile.id %}">Edit Profile</a>
+                          <a class="btn btn-info " href="{% url 'users:update_profile' user.profile.id %}">Edit Profile</a>
                         </div>
                       </div>
                     </div>
@@ -175,28 +175,28 @@
         {% endif %}
     {% elif update %}
         <h2>Edit Profile {{ user }}</h2><br>
-            <form method="post" action="/accounts/profile/">
-                {% csrf_token %}
-                <div class="form-row">
-                  <div class="form-group col-md-6 mb-0">
+        <form method="post">
+            {% csrf_token %}
+            <div class="form-row">
+                <div class="form-group col-md-6 mb-0">
                     {{ form.first_name|as_crispy_field }}
-                  </div>
-                  <div class="form-group col-md-6 mb-0">
+                </div>
+                <div class="form-group col-md-6 mb-0">
                     {{ form.last_name|as_crispy_field }}
-                  </div>
                 </div>
-                <div class="form-row">
-                  <div class="form-group col-md-6 mb-0">
+            </div>
+            <div class="form-row">
+                <div class="form-group col-md-6 mb-0">
                     {{ form.phone|as_crispy_field }}
-                  </div>
-                  <div class="form-group col-md-6 mb-0">
-                    {{ form.gender|as_crispy_field }}
-                  </div>
                 </div>
-                {{ form.about|as_crispy_field }}
-                {{ form.address|as_crispy_field }}
-                <button type="submit" class="btn btn-primary">Update</button>
-            </form>
-            <br>
+                <div class="form-group col-md-6 mb-0">
+                    {{ form.gender|as_crispy_field }}
+                </div>
+            </div>
+            {{ form.about|as_crispy_field }}
+            {{ form.address|as_crispy_field }}
+            <button type="submit" class="btn btn-primary">Update</button>
+        </form>
+        <br>
     {% endif %}
 {% endblock content %}
